lightningd/json: Move json helpers specific for lightningd to new module.
